requirements.txt
python-dotenv
streamlit
langchain_community
